Implement basic lobby for multiplayer launchgamemenu
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@3702 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/wlapplication.cc
@@ -1450,7 +1450,7 @@ struct SinglePlayerGameSettingsProvider : public GameSettingsProvider {
 		s.players[number] = ps;
 	}
 
-	virtual void setPlayerNumber(uint8_t number) {
+	virtual void setPlayerNumber(int32_t number) {
 		if (number >= s.players.size())
 			return;
